# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tickets Service API Documentation",
        "description": "L5 Swagger OpenApi API Documentation for Tickets Microservice",
        "contact": {
            "email": "bayusamudera@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Tickets Service API Server"
        }
    ],
    "paths": {
        "/api/v1/tickets": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Get list of tickets (Collection)",
                "operationId": "b406702bcd80642f4b5c83c1bf20105b",
                "responses": {
                    "200": {
                        "description": "List of tickets retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid Token or X-IAE-KEY"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Create a new ticket (Action)",
                "operationId": "80fbef860b264eed3bc3ccd90d8ddc8b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "schedule_id",
                                    "seat_number"
                                ],
                                "properties": {
                                    "schedule_id": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "seat_number": {
                                        "type": "string",
                                        "example": "A12"
                                    },
                                    "total_price": {
                                        "type": "integer",
                                        "example": 75000,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Resource created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/tickets/{id}": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Get details of a specific ticket (Resource)",
                "operationId": "596302276f442d583400ce2abf22cb00",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID of the ticket",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Resource not found"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Masukkan API Key (NIM: 102022400251)",
                "name": "X-IAE-KEY",
                "in": "header"
            },
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan Token JWT SSO untuk mengakses endpoint",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Tickets",
            "description": "Tickets"
        }
    ]
}